requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
concurrent-log-handler==0.9.12
Django==3.1
django-cors-headers==3.4.0
django-redis==4.12.1
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
hiredis==1.1.0
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
Naked==0.1.31
pycparser==2.20
pycryptodome==3.9.8
PyMySQL==0.10.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
shellescape==3.8.1
single-beat==0.4.2
six==1.15.0
sqlparse==0.3.1
tornadis==0.8.1
tornado==5.1.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
